ToolsUI de-clutter. Split off OpPanels, part 3.
