+++ ui/src/main/java/ucar/nc2/ui/GetDataRunnable.java
@@ -10,7 +10,7 @@ import java.io.IOException;
 /**
  *
  */
-interface GetDataRunnable {
+public interface GetDataRunnable {
     void run(Object o) throws IOException;
 }
