+++ ui/src/main/java/ucar/nc2/ui/op/DatasetTreeView.java
@@ -3,7 +3,7 @@
  * See LICENSE for license information.
  */
 
-package ucar.nc2.ui;
+package ucar.nc2.ui.op;
 
 import ucar.nc2.Dimension;
 import ucar.nc2.*;
