+++ ui/src/main/java/ucar/nc2/ui/op/FmrcCollectionPanel.java
@@ -6,36 +6,42 @@
 package ucar.nc2.ui.op;
 
 import ucar.nc2.ui.OpPanel;
+import ucar.nc2.ui.ToolsUI;
 import ucar.nc2.ui.widget.BAMutil;
 import ucar.util.prefs.PreferencesExt;
 
 import java.awt.BorderLayout;
+import java.beans.PropertyChangeEvent;
+import java.beans.PropertyChangeListener;
+import java.io.FileNotFoundException;
+import java.io.IOException;
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.util.Formatter;
 import javax.swing.AbstractButton;
+import javax.swing.JOptionPane;
+
+/**
+ * new Fmrc Collection Metadata storage in bdb
+ */
+public class FmrcCollectionPanel extends OpPanel {
+    private FmrcCollectionTable table;
 
 /**
  *
  */
-public class CollectionSpecPanel extends OpPanel {
-    private CollectionSpecTable table;
-
-/**
- *
- */
-    public CollectionSpecPanel(PreferencesExt dbPrefs) {
-        super(dbPrefs, "collection spec:", true, false);
-        table = new CollectionSpecTable(prefs);
+    public FmrcCollectionPanel(PreferencesExt dbPrefs) {
+        super(dbPrefs, "collection:", true, false);
+        table = new FmrcCollectionTable(prefs);
         add(table, BorderLayout.CENTER);
 
         final AbstractButton infoButton = BAMutil.makeButtcon("Information", "Detail Info", false);
         infoButton.addActionListener(e -> {
             Formatter f = new Formatter();
             try {
-                table.showCollection(f);
+                table.showInfo(f);
             }
-            catch (Exception e1) {
+            catch (IOException e1) {
                 final StringWriter sw = new StringWriter(5000);
                 e1.printStackTrace(new PrintWriter(sw));
                 f.format("%s", sw.toString());
@@ -45,18 +51,35 @@ public class CollectionSpecPanel extends OpPanel {
             detailWindow.show();
         });
         buttPanel.add(infoButton);
+
+        final AbstractButton refreshButton = BAMutil.makeButtcon("Undo", "Refresh", false);
+        refreshButton.addActionListener(e -> {
+            try {
+                table.refresh();
+            }
+            catch (Exception e1) {
+                final Formatter f = new Formatter();
+                final StringWriter sw = new StringWriter(5000);
+                e1.printStackTrace(new PrintWriter(sw));
+                f.format("%s", sw.toString());
+                detailTA.setText(f.toString());
+                detailTA.gotoTop();
+                detailWindow.show();
+            }
+        });
+        buttPanel.add(refreshButton);
     }
 
 /** */
     @Override
     public boolean process(Object o) {
-        final String command = (String) o;
+        String command = (String) o;
         if (command == null) {
             return false;
         }
 
         try {
-            table.setCollection(command);
+            // table.setCacheRoot(command);
             return true;
         }
         catch (Exception ioe) {
