+++ ui/src/main/java/ucar/nc2/ui/op/NCdumpPane.java
@@ -1,9 +1,9 @@
 /*
- * Copyright (c) 1998-2018 University Corporation for Atmospheric Research/Unidata
+ * Copyright (c) 1998-2019 University Corporation for Atmospheric Research/Unidata
  * See LICENSE for license information.
  */
 
-package ucar.nc2.ui;
+package ucar.nc2.ui.op;
 
 import ucar.ma2.InvalidRangeException;
 import ucar.ma2.IsMissingEvaluator;
