+++ ui/src/main/java/ucar/nc2/ui/op/NcStreamPanel.java
@@ -2,7 +2,7 @@
  * Copyright (c) 1998-2018 University Corporation for Atmospheric Research/Unidata
  * See LICENSE for license information.
  */
-package ucar.nc2.ui;
+package ucar.nc2.ui.op;
 
 import ucar.nc2.NetcdfFile;
 import ucar.nc2.NetcdfFileSubclass;
